+++ include/fix_parser.h
@@ -18,7 +18,6 @@ extern "C"
 
 FIXParser* fix_parser_create(char const* protFile, FIXParserAttrs const* attrs, int32_t flags);
 void fix_parser_free(FIXParser* parser);
-int32_t fix_parser_get_flags(FIXParser* parser);
 FIXMsg* fix_parser_fix_to_msg(FIXParser* parser, char const* data, uint32_t len, char delimiter, char const** stop);
 
 #ifdef __cplusplus
