+++ src/fix_msg_priv.h
@@ -40,6 +40,10 @@ int32_t fix_msg_del_field(FIXMsg* msg, FIXGroup* grp, uint32_t tag);
 FIXGroup* fix_msg_alloc_group(FIXMsg* msg);
 void fix_msg_free_group(FIXMsg* msg, FIXGroup* grp);
 
+int32_t fix_groups_to_string(FIXMsg* msg, FIXField* field, FIXFieldDescr* fdescr, char delimiter, char** buff, uint32_t* buffLen);
+int32_t int32_to_fix_msg(FIXParser* parser, uint32_t tag, int32_t val, char delimiter, uint32_t width, char padSym, char** buff, uint32_t* buffLen);
+int32_t fix_field_to_fix_msg(FIXParser* parser, FIXField* field, char delimiter, char** buff, uint32_t* buffLen);
+
 #ifdef __cplusplus
 }
 #endif
