+++ src/fix_parser.c
@@ -19,12 +19,6 @@
 
 #define CRC_FIELD_LEN 7
 
-static int validate_attrs(FIXParserAttrs* attrs);
-static int check_value(FIXFieldDescr* fdescr, char const* dbegin, char const* dend, char delimiter);
-static int32_t parse_group(FIXMsg* msg, int64_t numGroups, char const* data, uint32_t len, char delimiter, char const** stop);
-
-/*------------------------------------------------------------------------------------------------------------------------*/
-/* PUBLIC                                                                                                                 */
 /*------------------------------------------------------------------------------------------------------------------------*/
 FIXParser* fix_parser_create(char const* protFile, FIXParserAttrs const* attrs, int32_t flags)
 {
@@ -34,7 +28,7 @@ FIXParser* fix_parser_create(char const* protFile, FIXParserAttrs const* attrs, 
    {
       memcpy(&myattrs, attrs, sizeof(myattrs));
    }
-   if (!validate_attrs(&myattrs))
+   if (!fix_parser_validate_attrs(&myattrs))
    {
       return NULL;
    }
@@ -90,87 +84,6 @@ void fix_parser_free(FIXParser* parser)
 }
 
 /*------------------------------------------------------------------------------------------------------------------------*/
-FIXPage* fix_parser_alloc_page(FIXParser* parser, uint32_t pageSize)
-{
-   if (parser->attrs.maxPages > 0 && parser->attrs.maxPages == parser->used_pages)
-   {
-      fix_error_set(&parser->error,
-         FIX_ERROR_NO_MORE_PAGES, "No more pages available. MaxPages = %d, UsedPages = %d", parser->attrs.maxPages, parser->used_pages);
-      return NULL;
-   }
-   FIXPage* page = NULL;
-   if (parser->page == NULL) // no more free pages
-   {
-      uint32_t psize = (parser->attrs.pageSize > pageSize ? parser->attrs.pageSize : pageSize);
-      if (parser->attrs.maxPageSize > 0 && psize > parser->attrs.maxPageSize)
-      {
-         fix_error_set(
-               &parser->error,
-               FIX_ERROR_TOO_BIG_PAGE, "Requested new page is too big. MaxPageSize = %d, RequestedPageSize = %d",
-               parser->attrs.maxPageSize, psize);
-         return NULL;
-      }
-      page = calloc(1, sizeof(FIXPage) + psize - 1);
-      page->size = psize;
-   }
-   else
-   {
-      page = parser->page;
-      parser->page = page->next;
-      page->next = NULL; // detach from pool of free pages
-   }
-   ++parser->used_pages;
-   return page;
-}
-
-/*------------------------------------------------------------------------------------------------------------------------*/
-FIXPage* fix_parser_free_page(FIXParser* parser, FIXPage* page)
-{
-   FIXPage* next = page->next;
-   page->offset = 0;
-   page->next = parser->page;
-   parser->page = page;
-   --parser->used_pages;
-   return next;
-}
-
-/*------------------------------------------------------------------------------------------------------------------------*/
-FIXGroup* fix_parser_alloc_group(FIXParser* parser)
-{
-   if (parser->attrs.maxGroups > 0 && parser->attrs.maxGroups == parser->used_groups)
-   {
-      fix_error_set(&parser->error,
-         FIX_ERROR_NO_MORE_GROUPS,
-         "No more groups available. MaxGroups = %d, UsedGroups = %d", parser->attrs.maxGroups, parser->used_groups);
-      return NULL;
-   }
-   FIXGroup* group = NULL;
-   if (parser->group == NULL) // no more free group
-   {
-      group = calloc(1, sizeof(FIXGroup));
-   }
-   else
-   {
-      group = parser->group;
-      parser->group = group->next;
-      group->next = NULL; // detach from pool
-   }
-   ++parser->used_groups;
-   return group;
-}
-
-/*------------------------------------------------------------------------------------------------------------------------*/
-FIXGroup* fix_parser_free_group(FIXParser* parser, FIXGroup* group)
-{
-   FIXGroup* next = group->next;
-   memset(group, 0, sizeof(FIXGroup));
-   group->next = parser->group;
-   parser->group = group;
-   --parser->used_groups;
-   return next;
-}
-
-/*------------------------------------------------------------------------------------------------------------------------*/
 FIXMsg* fix_parser_fix_to_msg(FIXParser* parser, char const* data, uint32_t len, char delimiter, char const** stop)
 {
    if (!parser || !data)
@@ -181,7 +94,7 @@ FIXMsg* fix_parser_fix_to_msg(FIXParser* parser, char const* data, uint32_t len,
    int64_t tag = 0;
    char const* dbegin = NULL;
    char const* dend = NULL;
-   tag = parse_field(parser, data, len, delimiter, &dbegin, &dend);
+   tag = fix_parser_parse_field(parser, data, len, delimiter, &dbegin, &dend);
    if (tag == FIX_FAILED)
    {
       fix_error_set(&parser->error, FIX_ERROR_PARSE_MSG, "Unable to parse BeginString field.");
@@ -203,7 +116,7 @@ FIXMsg* fix_parser_fix_to_msg(FIXParser* parser, char const* data, uint32_t len,
       free(actualVer);
       return NULL;
    }
-   tag = parse_field(parser, dend + 1, len - (dend - data - 1), delimiter, &dbegin, &dend);
+   tag = fix_parser_parse_field(parser, dend + 1, len - (dend - data - 1), delimiter, &dbegin, &dend);
    if (tag == FIX_FAILED)
    {
       fix_error_set(&parser->error, FIX_ERROR_PARSE_MSG, "Unable to parse BodyLength field.");
@@ -226,7 +139,7 @@ FIXMsg* fix_parser_fix_to_msg(FIXParser* parser, char const* data, uint32_t len,
    }
    char const* crcbeg = NULL;
    char const* bodyEnd = dend + bodyLen;
-   tag = parse_field(parser, bodyEnd + 1, CRC_FIELD_LEN, delimiter, &crcbeg, stop);
+   tag = fix_parser_parse_field(parser, bodyEnd + 1, CRC_FIELD_LEN, delimiter, &crcbeg, stop);
    if (tag == FIX_FAILED)
    {
       fix_error_set(&parser->error, FIX_ERROR_PARSE_MSG, "Unable to parse CrcSum field.");
@@ -259,7 +172,7 @@ FIXMsg* fix_parser_fix_to_msg(FIXParser* parser, char const* data, uint32_t len,
          return NULL;
       }
    }
-   tag = parse_field(parser, dend + 1, bodyEnd - dend - 1, delimiter, &dbegin, &dend);
+   tag = fix_parser_parse_field(parser, dend + 1, bodyEnd - dend - 1, delimiter, &dbegin, &dend);
    if (tag == FIX_FAILED)
    {
       fix_error_set(&parser->error, FIX_ERROR_PARSE_MSG, "Unable to parse MsgType field.");
@@ -281,7 +194,7 @@ FIXMsg* fix_parser_fix_to_msg(FIXParser* parser, char const* data, uint32_t len,
    free(msgType);
    while(dend != bodyEnd)
    {
-      tag = parse_field(parser, dend + 1, bodyEnd - dend, delimiter, &dbegin, &dend);
+      tag = fix_parser_parse_field(parser, dend + 1, bodyEnd - dend, delimiter, &dbegin, &dend);
       if (tag == FIX_FAILED)
       {
          fix_msg_free(msg);
@@ -302,7 +215,7 @@ FIXMsg* fix_parser_fix_to_msg(FIXParser* parser, char const* data, uint32_t len,
       {
          if (parser->flags & PARSER_FLAG_CHECK_VALUE)
          {
-            if (check_value(fdescr, dbegin, dend, delimiter) == FIX_FAILED)
+            if (fix_parser_check_value(fdescr, dbegin, dend, delimiter) == FIX_FAILED)
             {
                fix_msg_free(msg);
                return NULL;
@@ -319,7 +232,7 @@ FIXMsg* fix_parser_fix_to_msg(FIXParser* parser, char const* data, uint32_t len,
                fix_error_set(&parser->error, FIX_ERROR_INVALID_ARGUMENT, "Unable to get group tag %d value.", tag);
                return NULL;
             }
-            res = parse_group(msg, num, dend + 1, bodyEnd - dend, delimiter, &dend);
+            res = fix_parser_parse_group(msg, num, dend + 1, bodyEnd - dend, delimiter, &dend);
             if (res == FIX_FAILED)
             {
                fix_msg_free(msg);
@@ -330,96 +243,4 @@ FIXMsg* fix_parser_fix_to_msg(FIXParser* parser, char const* data, uint32_t len,
       }
    }
    return msg;
-}
-
-/*------------------------------------------------------------------------------------------------------------------------*/
-/* PRIVATE                                                                                                                */
-/*------------------------------------------------------------------------------------------------------------------------*/
-static int validate_attrs(FIXParserAttrs* attrs)
-{
-   if (!attrs->pageSize)
-   {
-      attrs->pageSize = 4096;
-   }
-   if (!attrs->numPages)
-   {
-      attrs->numPages = 1000;
-   }
-   if (!attrs->numGroups)
-   {
-      attrs->numGroups = 1000;
-   }
-   if (attrs->maxPageSize > 0 && attrs->maxPageSize < attrs->pageSize)
-   {
-      fix_static_error_set(FIX_ERROR_INVALID_ARGUMENT, "ERROR: Parser attbutes are invalid: MaxPageSize < PageSize.");
-      return 0;
-   }
-   if (attrs->maxPages > 0 && attrs->maxPages < attrs->numPages)
-   {
-      fix_static_error_set(FIX_ERROR_INVALID_ARGUMENT, "Parser attbutes are invalid: MaxPages < NumPages.");
-      return 0;
-   }
-   if (attrs->maxGroups > 0 && attrs->maxGroups < attrs->numGroups)
-   {
-      fix_static_error_set(FIX_ERROR_INVALID_ARGUMENT, "Parser attbutes are invalid: MaxGroups < NumGroups.");
-      return 0;
-   }
-   return 1;
-}
-
-/*------------------------------------------------------------------------------------------------------------------------*/
-static int check_value(FIXFieldDescr* fdescr, char const* dbegin, char const* dend, char delimiter)
-{
-   if (IS_INT_TYPE(fdescr->type->valueType))
-   {
-      int64_t res = 0;
-      return fix_utils_atoi64(dbegin, dend - dbegin, delimiter, &res);
-   }
-   else if (IS_FLOAT_TYPE(fdescr->type->valueType))
-   {
-      double res = 0.0;
-      return fix_utils_atod(dbegin, dend - dbegin, delimiter, &res);
-   }
-   else if (IS_CHAR_TYPE(fdescr->type->valueType))
-   {
-      return (dend - dbegin == 1) ? FIX_SUCCESS : FIX_FAILED;
-   }
-   return FIX_SUCCESS;
-}
-
-/*------------------------------------------------------------------------------------------------------------------------*/
-int64_t parse_field(FIXParser* parser, char const* data, uint32_t len, char delimiter, char const** dbegin, char const** dend)
-{
-   int64_t num = 0;
-   int32_t res = fix_utils_atoi64(data, len, '=', &num);
-   if (res == FIX_FAILED)
-   {
-      fix_error_set(&parser->error, FIX_ERROR_INVALID_ARGUMENT, "Unable to extract field number.");
-      return FIX_FAILED;
-   }
-   len -= res;
-   *dend = *dbegin = data + res + 1;
-   for(;len > 0; --len)
-   {
-      if (**dend == delimiter)
-      {
-         break;
-      }
-      else
-      {
-         ++(*dend);
-      }
-   }
-   if (!len)
-   {
-      fix_error_set(&parser->error, FIX_ERROR_INVALID_ARGUMENT, "Field value must be terminated with '%c' delimiter.", delimiter);
-      return FIX_FAILED;
-   }
-   return num;
-}
-
-/*------------------------------------------------------------------------------------------------------------------------*/
-static int32_t parse_group(FIXMsg* msg, int64_t numGroups, char const* data, uint32_t len, char delimiter, char const** stop)
-{
-   return FIX_SUCCESS;
 }
